+++ packages/api/lambdas/executeMigrations.js
@@ -35,10 +35,6 @@ function handler(event, context, cb) {
     ],
     elasticsearch_host: process.env.ES_HOST,
 
-    // Used by migration2
-    granulesTable: process.env.GranulesTable,
-    filesTable: process.env.FilesTable,
-
     // Used by Migration_3
     bucket: process.env.system_bucket,
     stackName: process.env.stackName